feat: add generation ID
