+++ src/index.html
@@ -87,7 +87,7 @@
             <div class="key space lage">space</div>
           </ul>
         </div>
-      <div class="keyboard__info info">
+        <div class="keyboard__info info">
           <h2 class="info__title">Info</h2>
           <ul class="info__wrapper">
             <li class="info__language">
@@ -106,10 +106,10 @@
               <div class="key">Shift</div><span>Сочетание клавишь</span>
             </li>
           </ul>
-        </div> 
+        </div>
       </div>
     </div>
-  </section>  -->
+  </section> -->
 
 </body>
