+++ src/index.js
@@ -1,485 +1,11 @@
 import './main.scss';
+import buttons from './dataButtons';
 
-const buttons = [
-  [
-    {
-      values: {
-        en: {
-          lowercase: '`',
-          uppercase: '~',
-        },
-        ru: 'ё',
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '1',
-          uppercase: '!',
-        },
-        ru: {
-          lowercase: '1',
-          uppercase: '!',
-        },
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '2',
-          uppercase: '@',
-        },
-        ru: {
-          lowercase: '2',
-          uppercase: '"',
-        },
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '3',
-          uppercase: '#',
-        },
-        ru: {
-          lowercase: '3',
-          uppercase: '№',
-        },
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '4',
-          uppercase: '$',
-        },
-        ru: {
-          lowercase: '4',
-          uppercase: ';',
-        },
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '5',
-          uppercase: '%',
-        },
-        ru: {
-          lowercase: '5',
-          uppercase: '%',
-        },
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '6',
-          uppercase: '^',
-        },
-        ru: {
-          lowercase: '6',
-          uppercase: ': ',
-        },
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '7',
-          uppercase: '&',
-        },
-        ru: {
-          lowercase: '7',
-          uppercase: '?',
-        },
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '8',
-          uppercase: '*',
-        },
-        ru: {
-          lowercase: '8',
-          uppercase: '*',
-        },
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '9',
-          uppercase: '(',
-        },
-        ru: {
-          lowercase: '9',
-          uppercase: '(',
-        },
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '0',
-          uppercase: ')',
-        },
-        ru: {
-          lowercase: '0',
-          uppercase: ')',
-        },
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '-',
-          uppercase: '_',
-        },
-        ru: {
-          lowercase: '-',
-          uppercase: '_',
-        },
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '=',
-          uppercase: '+',
-        },
-        ru: {
-          lowercase: '=',
-          uppercase: '+',
-        },
-      },
-    },
-    {
-      values: 'backspace',
-      key: 'Backspace',
-    },
-  ],
-  [
-    {
-      values: 'tab',
-      key: 'Tab',
-    },
-    {
-      values: {
-        en: 'q',
-        ru: 'й',
-      },
-    },
-    {
-      values: {
-        en: 'w',
-        ru: 'ц',
-      },
-    },
-    {
-      values: {
-        en: 'e',
-        ru: 'у',
-      },
-    },
-    {
-      values: {
-        en: 'r',
-        ru: 'к',
-      },
-    },
-    {
-      values: {
-        en: 't',
-        ru: 'е',
-      },
-    },
-    {
-      values: {
-        en: 'y',
-        ru: 'н',
-      },
-    },
-    {
-      values: {
-        en: 'u',
-        ru: 'г',
-      },
-    },
-    {
-      values: {
-        en: 'i',
-        ru: 'ш',
-      },
-    },
-    {
-      values: {
-        en: 'o',
-        ru: 'щ',
-      },
-    },
-    {
-      values: {
-        en: 'p',
-        ru: 'з',
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '[',
-          uppercase: '{',
-        },
-        ru: 'х',
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: ']',
-          uppercase: '}',
-        },
-        ru: 'ъ',
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '\\',
-          uppercase: '|',
-        },
-        ru: {
-          lowercase: '/',
-          uppercase: '\\',
-        },
-      },
-    },
-    {
-      values: 'del',
-      key: 'Delete',
-    },
-  ],
-  [
-    {
-      values: 'caps',
-      key: 'Capslock',
-    },
-    {
-      values: {
-        en: 'a',
-        ru: 'ф',
-      },
-    },
-    {
-      values: {
-        en: 's',
-        ru: 'ы',
-      },
-    },
-    {
-      values: {
-        en: 'd',
-        ru: 'в',
-      },
-    },
-    {
-      values: {
-        en: 'f',
-        ru: 'а',
-      },
-    },
-    {
-      values: {
-        en: 'g',
-        ru: 'п',
-      },
-    },
-    {
-      values: {
-        en: 'h',
-        ru: 'р',
-      },
-    },
-    {
-      values: {
-        en: 'j',
-        ru: 'о',
-      },
-    },
-    {
-      values: {
-        en: 'k',
-        ru: 'л',
-      },
-    },
-    {
-      values: {
-        en: 'l',
-        ru: 'д',
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: ';',
-          uppercase: ':',
-        },
-        ru: 'ж',
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: "'",
-          uppercase: '"',
-        },
-        ru: 'э',
-      },
-    },
-    {
-      values: 'enter',
-      key: 'Enter',
-    },
-  ],
-  [
-    {
-      values: 'shift',
-      key: 'Shift',
-      action: function (keyboard) {
-        keyboard.switchMode();
-      },
-    },
-    {
-      values: {
-        en: 'z',
-        ru: 'я',
-      },
-    },
-    {
-      values: {
-        en: 'x',
-        ru: 'ч',
-      },
-    },
-    {
-      values: {
-        en: 'c',
-        ru: 'с',
-      },
-    },
-    {
-      values: {
-        en: 'v',
-        ru: 'м',
-      },
-    },
-    {
-      values: {
-        en: 'b',
-        ru: 'и',
-      },
-    },
-    {
-      values: {
-        en: 'n',
-        ru: 'т',
-      },
-    },
-    {
-      values: {
-        en: 'm',
-        ru: 'ь',
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: ',',
-          uppercase: '<',
-        },
-        ru: 'б',
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '.',
-          uppercase: '>',
-        },
-        ru: 'ю',
-      },
-    },
-    {
-      values: {
-        en: {
-          lowercase: '/',
-          uppercase: '?',
-        },
-        ru: {
-          lowercase: '.',
-          uppercase: ',',
-        },
-      },
-    },
-    {
-      values: '⇧',
-      key: 'ArrowUp',
-    },
-    {
-      values: 'shift',
-      key: 'Shift',
-      action: function (keyboard) {
-        keyboard.switchMode();
-      },
-    },
-  ],
-  [
-    {
-      values: 'ctrl',
-      key: 'Control',
-    },
-    {
-      values: 'win',
-      key: 'Meta',
-    },
-    {
-      values: 'alt',
-      key: 'Alt',
-    },
-    {
-      values: 'space',
-      key: ' ',
-      action: function () {},
-    },
-    {
-      values: 'alt',
-      key: 'Alt',
-    },
-
-    {
-      values: '⇦',
-      key: 'ArrowLeft',
-    },
-    {
-      values: '⇩',
-      key: 'ArrowDown',
-    },
-    {
-      values: '⇨',
-      key: 'ArrowRight',
-    },
-    {
-      values: 'ctrl',
-      key: 'Control',
-    },
-  ],
-];
 const appContainer = document.getElementById('app-keyboard');
 const toggleButton = document.getElementById('toggle-lang');
 
 const Keyboard = {
-  mode: 'uppercase',
+  mode: 'lowercase',
   locale: 'en',
 
   init() {
@@ -521,8 +47,22 @@ const Keyboard = {
     const textarea = document.createElement('textarea');
     textarea.id = 'text-display';
     textarea.classList.add('keyboard__text-display');
+    textarea.readOnly = true;
     return textarea;
   },
+
+  // {
+  //   values: {
+  //     en: {
+  //       lowercase: '1',
+  //       uppercase: '!',
+  //     },
+  //     ru: {
+  //       lowercase: '1',
+  //       uppercase: '!',
+  //     },
+  //   },
+  // },
 
   createKeys() {
     const keys = document.createElement('div');
@@ -536,6 +76,7 @@ const Keyboard = {
       rowButton.forEach((button) => {
         const buttonElement = document.createElement('li');
         buttonElement.classList.add('key');
+        buttonElement.id = button.id;
         if (typeof button.values === 'object') {
           if (typeof button.values[this.locale] === 'object') {
             buttonElement.textContent = button.values[this.locale][this.mode];
@@ -550,21 +91,46 @@ const Keyboard = {
       keys.append(buttonRowElement);
     });
 
-    keys.addEventListener('click', (event) => {
-      if (event.target.classList.contains('key')) {
-        const pressedKey = event.key;
-        const key = event.target.textContent;
-        console.log('Pressed key:', pressedKey);
-
-        // Действия при нажатии на клавишу
-        const value = this.getValue(
-          buttons.find((button) => button.values === key)
-        );
-        console.log('Value:', value);
+    keys.addEventListener('click', (e) => {
+      if (e.target.classList.contains('key')) {
+        const textArea = document.getElementById('text-display');
+        const buttonText = e.target.textContent;
+        textArea.value += buttonText;
       }
     });
 
+    document.addEventListener('keydown', (event) => {
+      console.log('file: index.js:554 ~ keys.addEventListener ~ event:', event);
+
+      const pressedKey = event.key;
+      console.log('Pressed key:', pressedKey);
+      const textArea = document.getElementById('text-display');
+      console.log(
+        'file: index.js:106 ~ document.addEventListener ~ textArea:',
+        textArea
+      );
+      const buttonText = e.target.textContent;
+      textArea.value += buttonText;
+      // if (this.getValue(pressedKey)) {
+      //   console.log(getValue(pressedKey));
+      // }
+
+      // Действия при нажатии на клавишу
+      // const value = this.getValue(
+      //   buttons.find((button) => button.values === key)
+      // );
+      // console.log('Value:', value);
+    });
+
     return keys;
+  },
+  switchMode() {
+    this.mode = this.mode === 'uppercase' ? 'lowercase' : 'uppercase';
+  },
+
+  switchLanguage() {
+    this.locale = this.locale === 'en' ? 'ru' : 'en';
+    console.log(this.locale);
   },
 
   getValue(letter) {
@@ -581,36 +147,8 @@ const Keyboard = {
       }
     }
   },
-
-  getTitle(button) {
-    if (button.title !== undefined) {
-      return button.title;
-    } else {
-      return this.getValue(button);
-    }
-  },
-
-  switchMode() {
-    this.mode = this.mode === 'uppercase' ? 'lowercase' : 'uppercase';
-  },
-
-  switchLanguage() {
-    this.locale = this.locale === 'en' ? 'ru' : 'en';
-    console.log(this.locale);
-  }, 
 };
 
 window.addEventListener('DOMContentLoaded', function () {
   Keyboard.init();
 });
-
-// let space = document.querySelector('.space');
-// let textareaDiv = document.getElementById('text-display');
-// console.dir('file: index.js:590 ~ textareaDiv:', textareaDiv.value);
-// console.log('file: index.js:589 ~ space:', space);
-// space.addEventListener('click', () => {
-//   textareaDiv.value += ' ';
-// });
-// textareaDiv.addEventListener('input', () => {
-//   console.log(textareaDiv.value);
-// });
