+++ src/main.scss
@@ -10,35 +10,22 @@
   flex-direction: column;
   max-width: 1100px;
   margin: 0 auto;
-  height: 100vh;
+  height: auto;
   width: 100%;
 }
 
-.keyboard__wrapper {
-  // display: grid;
-  // grid-template-columns: repeat(6, 1fr);
-  // grid-template-rows: repeat(6, 1fr);
-  // grid-column-gap: 20px;
-  // grid-row-gap: 20px;
-}
+// .keyboard__wrapper {
+// display: grid;
+// grid-template-columns: repeat(6, 1fr);
+// grid-template-rows: repeat(6, 1fr);
+// grid-column-gap: 20px;
+// grid-row-gap: 20px;
+// }
 
 .keyboard__title {
   font-size: 3rem;
   text-align: center;
 }
-// .keyboard {
-//   display: flex;
-//   flex-direction: column;
-//   align-items: center;
-//   justify-content: center;
-//   width: 850px;
-//   height: 400px;
-//   padding: 40px 10px;
-//   border-radius: 15px;
-//   background: #676767;
-//   box-shadow: -4px 4px 9px #363636, 4px -4px 9px #989898;
-//   margin: 30px;
-// }
 
 .keyboard__container {
   display: flex;
@@ -64,17 +51,11 @@
 .keyboard__keys {
   width: 70%;
   width: 100%;
-  margin-top: 30px;
-  // display: flex;
-  //   flex-direction: column;
-  //   align-items: center;
-  //   justify-content: center;
-  //
-    padding: 40px 10px;
-    border-radius: 15px;
+  padding: 40px 10px;
+  border-radius: 15px;
   //   background: #676767;
-    box-shadow: -4px 4px 9px #363636, 4px -4px 9px #989898;
-  //   margin: 30px;
+  box-shadow: -4px 4px 9px #363636, 4px -4px 9px #989898;
+  margin: 30px 0;
 }
 
 .keyboard__row {
@@ -122,40 +103,45 @@
   background: #f4c240;
   box-shadow: inset -7px 7px 14px #b28e2f, inset 7px -7px 14px #fff651;
 }
+.key.active {
+  border-radius: 6px;
+  background: #3cbf3a;
+  box-shadow: inset -7px 7px 14px #2b8529, inset 7px -7px 14px #4bf048;
+}
 
 // Checkbox language
-.checkbox-wrapper input[type="checkbox"] {
+.checkbox-wrapper input[type='checkbox'] {
   background-size: 100% 100%, 200% 100%;
   background-position: 0 0, 15px 0;
   border-radius: 25px;
-  box-shadow: inset 0 1px 4px hsla(0,0%,0%,.5),
-                inset 0 0 10px hsla(0,0%,0%,.5),
-                0 0 0 1px hsla(0,0%,0%,.1),
-                0 -1px 2px 2px hsla(0,0%,0%,.25),
-                0 2px 2px 2px hsla(0,0%,100%,.75);
+  box-shadow: inset 0 1px 4px hsla(0, 0%, 0%, 0.5),
+    inset 0 0 10px hsla(0, 0%, 0%, 0.5), 0 0 0 1px hsla(0, 0%, 0%, 0.1),
+    0 -1px 2px 2px hsla(0, 0%, 0%, 0.25), 0 2px 2px 2px hsla(0, 0%, 100%, 0.75);
   cursor: pointer;
   height: 25px;
   padding-right: 25px;
   width: 75px;
   -webkit-appearance: none;
-  -webkit-transition: .25s;
+  -webkit-transition: 0.25s;
 }
 
-.checkbox-wrapper input[type="checkbox"]:after {
+.checkbox-wrapper input[type='checkbox']:after {
   background-color: #eee;
-  background-image: -webkit-linear-gradient(hsla(0,0%,100%,.1), hsla(0,0%,0%,.1));
+  background-image: -webkit-linear-gradient(
+    hsla(0, 0%, 100%, 0.1),
+    hsla(0, 0%, 0%, 0.1)
+  );
   border-radius: 25px;
-  box-shadow: inset 0 1px 1px 1px hsla(0,0%,100%,1),
-                inset 0 -1px 1px 1px hsla(0,0%,0%,.25),
-                0 1px 3px 1px hsla(0,0%,0%,.5),
-                0 0 2px hsla(0,0%,0%,.25);
+  box-shadow: inset 0 1px 1px 1px hsla(0, 0%, 100%, 1),
+    inset 0 -1px 1px 1px hsla(0, 0%, 0%, 0.25),
+    0 1px 3px 1px hsla(0, 0%, 0%, 0.5), 0 0 2px hsla(0, 0%, 0%, 0.25);
   content: '';
   display: block;
   height: 25px;
   width: 50px;
 }
 
-.checkbox-wrapper input[type="checkbox"]:checked {
+.checkbox-wrapper input[type='checkbox']:checked {
   background-position: 0 0, 35px 0;
   padding-left: 25px;
   padding-right: 0;
